Refactored contexts and examples in repo specs
Contexts moved to context/ and 'context' removed from file names.
'example' removed from example file names.
