+++ spec/fortnox/api/repositories/contexts/json_helper.rb
