+++ spec/fortnox/api/repositories/examples/all.rb
