+++ spec/fortnox/api/repositories/invoice_spec.rb
@@ -1,12 +1,12 @@
 require 'spec_helper'
-require 'fortnox/api/repositories/environment_context'
+require 'fortnox/api/repositories/contexts/environment'
 require 'fortnox/api/repositories/invoice'
-require 'fortnox/api/repositories/examples/all_examples'
-require 'fortnox/api/repositories/examples/find_examples'
-require 'fortnox/api/repositories/examples/save_examples'
+require 'fortnox/api/repositories/examples/all'
+require 'fortnox/api/repositories/examples/find'
+require 'fortnox/api/repositories/examples/save'
 
 describe Fortnox::API::Repository::Invoice do
-  include_context 'environment context'
+  include_context 'environment'
 
   include_examples '.all'
